+++ scripts/experiments/smoke_test_sam3.py
